requirements.txt
Python-IO
datetime
requests
openpyxl
certifi==2025.8.3
charset-normalizer==3.4.3
idna==3.10
numpy==2.2.6
pandas==2.3.2
python-dateutil==2.9.0.post0
pytz==2025.2
six==1.17.0
tzdata==2025.2
urllib3==2.5.0